Move Glasbey and LogHistogramDataset in suitable packages.
