+++ src/main/java/fiji/plugin/trackmate/gui/panels/FilterPanel.java
@@ -51,7 +51,6 @@ import org.jfree.chart.plot.PlotOrientation;
 import org.jfree.chart.plot.XYPlot;
 import org.jfree.chart.renderer.xy.StandardXYBarPainter;
 import org.jfree.chart.renderer.xy.XYBarRenderer;
-import org.jfree.data.statistics.LogHistogramDataset;
 
 import fiji.plugin.trackmate.features.FeatureFilter;
 import fiji.plugin.trackmate.features.spot.SpotContrastAndSNRAnalyzerFactory;
