+++ src/main/java/fiji/plugin/trackmate/visualization/PerTrackFeatureColorGenerator.java
@@ -8,7 +8,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-import org.jfree.chart.renderer.GlasbeyLut;
 import org.jfree.chart.renderer.InterpolatePaintScale;
 import org.jgrapht.graph.DefaultWeightedEdge;
